Keep survey-related JS in the surveys module
