+++ jsconfig.json
@@ -15,7 +15,6 @@
       "indico/modules/events/editing/*": ["./indico/modules/events/editing/client/js/*"],
       "indico/modules/events/management/*": ["./indico/modules/events/management/client/js/*"],
       "indico/modules/events/persons/*": ["./indico/modules/events/persons/client/js/*"],
-      "indico/modules/events/surveys/*": ["./indico/modules/events/surveys/client/js"],
       "indico/modules/events/*": ["./indico/modules/events/client/js/*"],
       "indico/*": ["./indico/web/client/js/*"],
       "axios/unsafe/*": ["./node_modules/axios/lib/*"]
